@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:scoped_model/scoped_model.dart';
 
 class Quote {
-  final String quoteSurah;
-  final String quoteText;
-  final String quoteReference;
+  final String? quoteSurah;
+  final String? quoteText;
+  final String? quoteReference;
   Quote({this.quoteReference, this.quoteSurah, this.quoteText});
 }
 
@@ -17,9 +17,9 @@ class QuoteModel extends Model {
   );
 
 //this get the properties of the Quote class set in _quote variable
-  String get quoteSurah => _quote.quoteSurah;
-  String get getquoteReference => _quote.quoteReference;
-  String get quoteText => _quote.quoteText;
+  String? get quoteSurah => _quote.quoteSurah;
+  String? get getquoteReference => _quote.quoteReference;
+  String? get quoteText => _quote.quoteText;
 
 //this function makes the model more readable for frequent acess
   static QuoteModel of(BuildContext context) =>
